network change without errors
