+++ src/app/[network]/(App)/Background.tsx
