+++ src/app/[network]/(App)/Navbar.tsx
