+++ src/app/[network]/(App)/bulkcore1/AnalyticSection.tsx
