+++ src/app/[network]/(App)/bulkcore1/BuyCores.tsx
