+++ src/app/[network]/(App)/bulkcore1/CoreOwners.tsx
