+++ src/app/[network]/(App)/bulkcore1/CoreUtilisation.tsx
