+++ src/app/[network]/(App)/bulkcore1/page.tsx
