+++ src/app/[network]/(App)/bulkcore2/AnalyticSection.tsx
