+++ src/app/[network]/(App)/bulkcore2/CoreOwners.tsx
