+++ src/app/[network]/(App)/bulkcore2/MiniGraphSection.tsx
