+++ src/app/[network]/(App)/bulkcore2/SearchSection.tsx
