+++ src/app/[network]/(App)/bulkcore2/page.tsx
