+++ src/app/[network]/(App)/core/[number]/[regionId]/[mask]/Core.tsx
