+++ src/app/[network]/(App)/core/[number]/[regionId]/[mask]/page.tsx
