+++ src/app/[network]/(App)/globals.css
