+++ src/app/[network]/(App)/instacore/AnalyticSection.tsx
