+++ src/app/[network]/(App)/instacore/InstaGraph.tsx
