+++ src/app/[network]/(App)/instacore/page.tsx
