+++ src/app/[network]/(App)/my-cores/AnalyticSection.tsx
