+++ src/app/[network]/(App)/my-cores/MyCores.tsx
