+++ src/app/[network]/(App)/my-cores/page.tsx
