+++ src/app/[network]/(App)/renewal/AnalyticSection.tsx
