+++ src/components/web3/SupportedChains.tsx
@@ -1,35 +1,25 @@
-'use client'
-
-import { env } from '@/config/environment'
-import { getCurrentChain } from '@/utils/common/getCurrentChain'
-import { SubstrateChain, getSubstrateChain, useInkathon } from '@poppyseed/lastic-sdk'
-import Cookies from 'js-cookie'
+import { SUPPORTED_CHAINS } from '@/config/environment'
+import { switchChainInPath } from '@/utils/common/chainPath'
+import { useInkathon } from '@poppyseed/lastic-sdk'
+import Link from 'next/link'
+import { usePathname } from 'next/navigation'
 import { FC, useState } from 'react'
-import { toast } from 'react-hot-toast'
 import { FiChevronDown } from 'react-icons/fi'
 
 const SupportedChains: FC = () => {
-  const { activeChain, switchActiveChain } = useInkathon()
-  const { defaultChain } = getCurrentChain()
+  const { activeChain } = useInkathon()
+  //const { activeChain, switchActiveChain } = useInkathon()
+  const pathname = usePathname()
 
   const [isDropdownOpen, setIsDropdownOpen] = useState(false)
 
-  const [supportedChains] = useState(
-    env.supportedChains.map((chainMap) => ({
-      coretime: getSubstrateChain(chainMap.coretime) as SubstrateChain,
-      relay: getSubstrateChain(chainMap.relay) as SubstrateChain,
-    })),
-  )
-
-  const handleChainSwitch = async (chain: SubstrateChain, relayChain: SubstrateChain) => {
-    if (chain.network !== activeChain?.network) {
-      Cookies.set('activeChain', chain.network, { expires: 7 }) // Store the active chain in cookies for 7 days
-      Cookies.set('activeRelayChain', relayChain.network, { expires: 7 }) // Store the active relay chian in cookies for 7 days
-      await switchActiveChain?.(chain, relayChain)
-      toast.success(`Switched to ${chain.name}`)
-      setIsDropdownOpen(false) // Close dropdown after switching
-    }
-  }
+  //const handleChainSwitch = async (chain: SubstrateChain, relayChain: SubstrateChain) => {
+  // if (chain.network !== activeChain?.network) {
+  //   await switchActiveChain?.(chain, relayChain)
+  //   toast.success(`Switched to ${chain.name}`)
+  //   setIsDropdownOpen(false) // Close dropdown after switching
+  // }
+  //}
 
   return (
     <div className="relative flex flex-row">
@@ -37,24 +27,25 @@ const SupportedChains: FC = () => {
         className="flex flex-row justify-between space-x-3 items-center p-2 cursor-pointer"
         onClick={() => setIsDropdownOpen((prev) => !prev)}
       >
-        <p>Network: {defaultChain}</p>
+        <p>Network: {activeChain?.name}</p>
         <FiChevronDown size={16} className={`${isDropdownOpen ? 'transform rotate-180' : ''}`} />
       </div>
 
       {/* Dropdown List */}
       {isDropdownOpen && (
         <div className="absolute top-full left-0 w-full border-t-2 border-gray-3">
-          {supportedChains.map((chainPair, index) => {
+          {SUPPORTED_CHAINS.map((chainPair, index) => {
             const { coretime, relay } = chainPair
             return (
-              coretime.network !== activeChain?.network && (
-                <div
+              coretime !== activeChain?.network && (
+                <Link
                   key={index}
-                  onClick={() => handleChainSwitch(coretime, relay)}
+                  //onClick={() => handleChainSwitch(coretime, relay)}
+                  href={switchChainInPath(pathname, relay)}
                   className="p-2 flex justify-between items-center cursor-pointer"
                 >
-                  <p>{coretime.name}</p>
-                </div>
+                  <p>{coretime}</p>
+                </Link>
               )
             )
           })}
