--- src/config/types.ts
@@ -1,4 +1,0 @@
-export type ChainConfig = {
-  coretime: string
-  relay: string
-}
